New fl_draw.h functions to take arrays of points to draw shapes, and to use double for all arguments. Not compete yet, I want to allow fonts to be chosen with floating point sizes.
Made the documentation use css. Nothing too complicated and it should
still go through htmldoc. Somebody with more design sense than me
should try to improve fltk.css.

Clean up of Fl_Valuator to get rid of a lot of confusing functions and
make linesize & pagesize be the actual increments.  Fl_Slider draws
tick marks at intelligent distances (I hope), and if step() is
non-zero it does a lot of work to select "nice" values (powers of 10
times 1,2, or 5) depending on the slider scale.


git-svn-id: file:///fltk/svn/trunk@2335 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/Fl_Input.h
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_Input.h,v 1.7 2002/06/21 06:17:09 spitzak Exp $"
+// "$Id: Fl_Input.h,v 1.8 2002/07/01 15:28:19 spitzak Exp $"
 //
 // Input base class header file for the Fast Light Tool Kit (FLTK).
 //
@@ -96,7 +96,7 @@ private:
   int label_width;
 
   const char* expand(const char*, char*, int) const;
-  int expandpos(const char*, const char*, const char*, int*) const;
+  double expandpos(const char*, const char*, const char*, int*) const;
   void minimal_update(int, int);
   void minimal_update(int p);
   void erase_cursor_at(int p);
@@ -121,5 +121,5 @@ private:
 #endif
 
 //
-// End of "$Id: Fl_Input.h,v 1.7 2002/06/21 06:17:09 spitzak Exp $".
+// End of "$Id: Fl_Input.h,v 1.8 2002/07/01 15:28:19 spitzak Exp $".
 //
